fix: Better templates
The templates for each lessons now mentions where to put the
code. Also some return values are already coded, to minimize
the boilerplate code that needs to be written.

Due to those changes, the code from the next lesson isn't a
full solution to the current lesson anymore. In order to have
the solution of a lesson somewhere, there's now a constant
called `_solution` which contains the full source of the
completed lesson.