feat: blog lessons (#14)
Notes:

 - Currently the pre-populated code is the solution of the previous
   lesson. So you can e.g. copy the code from lesson 3 into the code
   box of lesson 2 and it should validate correctly.
 - Validations for lesson 1 is done
 - Validations for lesson 2 are almost done (one is missing
 - Validations for other lessons is missing
 - I'm still unclear on how to validate lesson 6

* WIP: blog lessons2

Notes:

 - Currently the pre-populated code is the solution of the previous
   lesson. So you can e.g. copy the code from lesson 3 into the code
   box of lesson 2 and it should validate correctly.
 - Validations for lesson 1-5 is done
 - Validations for other lessons is missing
 - I'm still unclear on how to validate lesson 7

* fix: remove logging statement

* fix: add validation to blog lesson 6

* fix: fix runable code

* chore: add `shallow-equal` module to package-lock.json

* fix: improve blog lesson 7

It has slighly better instructions now and valiation

* fix: make order of links not matter anymore

Don't check the CIDs as the order of the links really shouldn't matter.

* fix: reduce number of tags in blog lessons

More tags don't add any value, hence restricting it to the minimum that
makes sense.

Also make the validation less strict. It often doesn't make sense to
check the CIDs as a totally valid solution still lead to different
CIDs.

* fix: remove findLinks()

The `findLinks()` function wasn't really helpful.

* fix: Better templates

The templates for each lessons now mentions where to put the
code. Also some return values are already coded, to minimize
the boilerplate code that needs to be written.

Due to those changes, the code from the next lesson isn't a
full solution to the current lesson anymore. In order to have
the solution of a lesson somewhere, there's now a constant
called `_solution` which contains the full source of the
completed lesson.

* fix: replace the permanent Web with a blog post

Remove the permanent Web lessons from the home screen and replace
them with the blog post lessons.