fix: update monaco as local dep. fix editor resize.
This is mostly voodoo. Monaco is not a big fan of being resized,
but with heights on the container divs we can get what we want
which is an explicit height set from the number of code lines
when the editor is inline, and a variable height that takes the
remaining vertical space when in full screen mode.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>