Provide issue pre-fills via link; update text
- Build the link to a new issue programmatically, including the lesson
name, number, and path in the issue title and some questions we hope the
user will answer in the body.
- Remove request for user to copy URL, since we now are providing this
automagically.
- TODO: Not sure how to match max width of lesson box, though the text
here is no longer wide enough for this to matter.
- TODO: Create a specific GitHub issue template to use for questions
originating from this lesson page, so that nicer formatting can be used
in the body. (Don't know how to create line breaks in the current
parameter format.)