From last lesson of series, link to workshop list
- Addresses issues #64 & #38, in which the the "Next" button appeared on
the last lesson in a series and directed the user to a non-existent
subsequent lesson. This is resolved by now showing instead a "More
Workshops" button that goes back to the main list of all workshops so
the user can choose what they'd like to do next.
- In `Lesson.vue`, change the conditionals for success buttons to
include an option (when the solution is correct and `last` has a value
of ``"true"`) that loads a button with the text "More Workshops," bound
to a new function `workshopMenu` that clears the output and directs the
user to the main page, rather than incrementing the lesson number.
- In Basics lesson 3 and Blog lesson 7 (the last of each series), follow
the steps listed below.
- WHEN CREATING THE LAST LESSON OF A NEW WORKSHOP:  Add `last="true"`
to the values being passed to the Lesson component. Use the success
message "Great job! You've completed this series of lessons!" or
something else that makes it clear this was the last lesson in the
series. The button will populate automatically.