feat: add "get help" link to github issues.
* Add link to file issue if confused by lesson

- Added link to GitHub repo for those who need help or have suggestions
to improve the lessons
- TODO: Improve styling

* Provide issue pre-fills via link; update text

- Build the link to a new issue programmatically, including the lesson
name, number, and path in the issue title and some questions we hope the
user will answer in the body.
- Remove request for user to copy URL, since we now are providing this
automagically.
- TODO: Not sure how to match max width of lesson box, though the text
here is no longer wide enough for this to matter.
- TODO: Create a specific GitHub issue template to use for questions
originating from this lesson page, so that nicer formatting can be used
in the body. (Don't know how to create line breaks in the current
parameter format.)

* Update caps and use template string

- Use template string for improved readability of issueUrl
- Add formatting to issue body
- Change `Url` to `url` and `imageURL` to `imageUrl` to meet standards

* Set max width on footnote section

- Set a max width on the section containing the issue link to ensure it
doesn't get wider than the exercise box should we later change its text.
- Use <p> tags around this wording.

* Move issueUrl composition to `data: self`

- Create the issueUrl in `data: self` instead of in `computed`
- Although it currently works, I can't figure out any way to access the
`lessonNumber` that lives in `computed`, nor to make a functional
`lessonNumber` live in `data: self` so I can pull it from there. I had
to rebuild the whole thing within the `issueUrl`, so it adds more code
to the template string.

* Correct grammar in success message

- Change success message from "All works!" to "Everything works!"
- Addresses issue #65

* Remove merge comments

- Remove comment lines from earlier merge

* From last lesson of series, link to workshop list

- Addresses issues #64 & #38, in which the the "Next" button appeared on
the last lesson in a series and directed the user to a non-existent
subsequent lesson. This is resolved by now showing instead a "More
Workshops" button that goes back to the main list of all workshops so
the user can choose what they'd like to do next.
- In `Lesson.vue`, change the conditionals for success buttons to
include an option (when the solution is correct and `last` has a value
of ``"true"`) that loads a button with the text "More Workshops," bound
to a new function `workshopMenu` that clears the output and directs the
user to the main page, rather than incrementing the lesson number.
- In Basics lesson 3 and Blog lesson 7 (the last of each series), follow
the steps listed below.
- WHEN CREATING THE LAST LESSON OF A NEW WORKSHOP:  Add `last="true"`
to the values being passed to the Lesson component. Use the success
message "Great job! You've completed this series of lessons!" or
something else that makes it clear this was the last lesson in the
series. The button will populate automatically.

* Correct typos in blog lesson 3

- change "ipld" to "ipfs" in `ipfs.dag.put()`
- change "array" to "an array"

* Open new GitHub issue in new browser window

- In `Lesson.vue`, add `target="_blank"` to link so that when a user
clicks the link to log an issue about a question they have, their
original window stays open and anything they've typed in the exercise
box isn't lost.

* Use backticks instead of double quotes for strings

- Use backticks for template string instead of double quotes when the
string needs to include an apostrophe in the success message for lessons
Basics 3 and Blog 7