Reset code to default when cache is cleared
- Refactor `clearCache` and `onCodeChange` functions to ensure that when
the user requests to clear the cache, the `this.cachedCode` is set to
`false`, `this.code` is set to the default for the lesson
(`this.$attrs.code`), and the content of the editor box is set to
display the new code (`this.editor.setValue(this.code)`).
- Functinos now check whether the code is equal to the default code
because otherwise when you reset to the default code, it sees it as a
code change and stores the default as cached, thereby making the label
appear that says you're looking at cached code when it's not true.
- With the new refactoring, the system also catches when you delete your
typed code and get yourself back to the default by yourself.
- Adjusted the message displayed so it makes sense for the live updates.
- Adjusted message formatting so it displays in green to make it stick
out from the rest of the instructional text.