Save lessonPassed status on successful submit
- Add `lessonKey` property to match the route path
- When code submitted is correct, create a localStorage entry with the
lessonKey as its key
- Add `lessonPassed` property that tracks whether that localStorage
entry is present (it sets itself correctly on first load but can't watch
for localStorage changes automatically, so its value is also updated
when the code is submitted, which affects the v-if success messages
- TODO: Display completed status from the homepage in list of lessons
(with a checkmark or something similar)
- TODO: Currently once you've passed the lesson it stays passed, and you
have to manually delete the localStorage entry from the command line and
refresh the page to go back. Offer the user an option to reset to
incomplete status (`delete localStorage[this.lessonKey]`)
- TODO: Currently if you return to a completed lesson the status of your
code may or may not match the right answer you used to pass. Decide how
we want the user flow to work. If you change the code away from the correct
answer should you no longer have passed? If you load correct cached code
should the submit button instead be the next lesson button?