fix: ensure ipfs browser bundle is used
- alias the ipfs dependency to the pre-built bundle.
- update link to explore.ipld.io
- update ipfs to latest release

This ensures the browser friendly ipfs bundle is used. Before this
PR we were seeing a bunch of errors when IPFS tried to init, which
are caused by trying to get webpack to bundle the node build and
not substituting all the browser friendly deps correctly. Luckily,
this work is done for us in the ipfs dist build, so just use that.

fixes #164
fixes #125

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>