fix: open mailchimp submission in new tab
this approaches disables the jsonp request and uses the good old form
action method.
this fixes the issue in firefox issue when content blocking is enabled.
the downside of this is that we can't know when the subscription fails,
and since we are always showing the success message now it might not
match what the user has seen in the mailchimp browser tab.