feat: add external link option for the button component
for this we had to split the button into two components: Button and
ButtonLink.

The normal Button component should be used for situations when we
need to control the button click action.
The ButtonLink is simply an anchor tag with the look and feel of the
button with the new option to be an external link.