fix: force page refresh when ChunkLoadError happens
https://github.com/ProtoSchool/protoschool.github.io/issues/555
This is cache issue.
When a user comes back to a tab with ProtoSchool open
after a deployment, they might get this ChunkLoadError error
because the chunk's name does not exist anymore
(because of cache busting hashes in the filenames).

the fix for this is to force the browser to reload.