fix: do not hide ThankYou message with link to profile survey
after the user finishes the tutorial feedback survey and the user
does not click the profile survey link, then we should not hide
the profile survey link when the user comes back to the page or
finishes another tutorial. it should still show it to again give
the opportunity to the user to do it.